requirements.txt
PyYAML
torch
torchtext

